# httpapi/openapi.yaml
openapi: "3.1.0"
info:
  title: Pit Officer REST API
  version: "1"
  description: >
    Control-plane REST API for the Pit Officer service. The public surface is
    mounted at `/api/v1`; a mirror at `/app/api/v1` is reserved for the operator
    dashboard and carries the same routes except the route manifest, which is
    published only on `/api/v1`. Control endpoints answer `application/json`;
    endpoints that hand back a file answer with that file's own media type.
    A malformed request body, or one that violates the schema or a parameter
    constraint, answers RFC 9457 `application/problem+json` - `400` for broken
    syntax, `422` for a constraint violation. Every other error answers the
    envelope `{"error":{"code":"...","message":"..."}}`.
    Money and size values
    are exact decimal strings - never IEEE 754 floats.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /api/v1
    description: Public REST API
...
paths:
...
  /accounts/{code}/adjustments:
    parameters:
      - $ref: "#/components/parameters/AccountCode"
    get:
      operationId: listAccountAdjustments
      summary: List adjustments for an account
      tags: [adjustments]
      parameters:
        - $ref: "#/components/parameters/SourceFilter"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Adjustment list.
          content:
            application/json:
              schema:
                type: object
                required: [adjustments]
                properties:
                  adjustments:
                    type: array
                    items:
                      $ref: "#/components/schemas/Adjustment"
        "400":
          $ref: "#/components/responses/MalformedRequestError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      operationId: applyAdjustment
      summary: Apply a balance adjustment to an account
      description: >
        Runs the adjustment through the engine. A policy reject is a successful
        call: the rejected record is returned in the body with `status` =
        `"rejected"`. HTTP errors indicate only system-level failures. The request
        may carry an optional `id`: when supplied it is used as-is and a
        duplicate is rejected with 409; when omitted the server generates one and
        returns it on the record. If the engine returns no account modifications,
        Officer records nothing and returns 204.
      tags: [adjustments]
      parameters:
        - $ref: "#/components/parameters/MissingAccount"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustmentRequest"
      responses:
        "201":
          description: Adjustment record created (accepted or rejected).
          content:
            application/json:
              schema:
                type: object
                required: [adjustment]
                properties:
                  adjustment:
                    $ref: "#/components/schemas/Adjustment"
        "204":
          description: The engine returned no account modifications.
        "400":
          $ref: "#/components/responses/MalformedRequestError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/AccountMissingOrNotFoundError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------------------------------------------------------------------------
  # Groups
  # ---------------------------------------------------------------------------
...
  /accounts/{code}/balances/realized-pnl:
    parameters:
      - $ref: "#/components/parameters/AccountCode"
    put:
      operationId: setBalanceRealizedPnl
      summary: Set realized P&L for a balance row
      description: >
        Stores the current realized P&L snapshot for one per-(account, asset)
        balance row as an Officer control-plane record. This endpoint does not
        call the SDK account adjustment API and does not seed the SpotFunds
        account-currency P&L kill-switch accumulator.
      tags: [balances]
      parameters:
        - $ref: "#/components/parameters/MissingAccount"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BalanceRealizedPnlRequest"
      responses:
        "200":
          description: Updated balance row.
          content:
            application/json:
              schema:
                type: object
                required: [balance]
                properties:
                  balance:
                    $ref: "#/components/schemas/Balance"
        "400":
          $ref: "#/components/responses/MalformedRequestError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/AccountMissingOrNotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------------------------------------------------------------------------
  # Balances
  # ---------------------------------------------------------------------------
...
  /adjustments:
    get:
      operationId: listAdjustments
      summary: List adjustments across all accounts
      tags: [adjustments]
      parameters:
        - name: id
          in: query
          description: Exact adjustment public ID lookup.
          schema:
            type: string
        - name: account
          in: query
          description: Exact account code.
          schema:
            type: string
        - name: asset
          in: query
          description: Exact adjusted asset code.
          schema:
            type: string
        - name: status
          in: query
          description: Filter by adjustment outcome status.
          schema:
            type: string
            enum: [all, accepted, rejected]
            default: all
        - name: atMode
          in: query
          schema:
            type: string
            enum: [all, gt, after, greater_than, gte, on_or_after, lt, before, less_than, lte, on_or_before, between]
            default: all
          description: Adjustment timestamp comparison mode.
        - name: atMin
          in: query
          schema:
            type: string
            format: date-time
          description: Lower adjustment timestamp bound.
        - name: atMax
          in: query
          schema:
            type: string
            format: date-time
          description: Upper adjustment timestamp bound.
        - name: sort
          in: query
          schema:
            type: string
            enum: [account, asset, at, principal, source, status]
          description: Adjustment list sort key. Unknown keys are rejected.
        - $ref: "#/components/parameters/SourceFilter"
        - $ref: "#/components/parameters/SortOrder"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/OffsetParam"
        - $ref: "#/components/parameters/PageParam"
      responses:
        "200":
          description: Adjustment list.
          content:
            application/json:
              schema:
                type: object
                required: [adjustments, total]
                properties:
                  adjustments:
                    type: array
                    items:
                      $ref: "#/components/schemas/Adjustment"
                  total:
                    type: integer
                    description: Total adjustments matching the filters before paging.
        "400":
          $ref: "#/components/responses/MalformedRequestError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------------------------------------------------------------------------
  # Orders
  # ---------------------------------------------------------------------------
...
components:
...
  parameters:
    AccountCode:
      name: code
      in: path
      required: true
      description: URL-encoded account public code.
      schema:
        type: string
...
    SourceFilter:
      name: source
      in: query
      description: "Filter by originating surface: `panel`, `api`, `mcp`, or `system`."
      schema:
        type: string
        enum: [panel, api, mcp, system]
    SortOrder:
      name: order
      in: query
      description: Sort direction.
      schema:
        type: string
        enum: [asc, desc]
        default: asc
...
    MissingAccount:
      name: missingAccount
      in: query
      required: true
      description: >
        Explicit choice for a request that names an account Officer does not
        know yet: `create` registers it with default settings (no group, no
        currency) and applies the request; `reject` fails with 404
        `account_missing`. There is no default - the value is required whenever
        the request names a non-empty account code.
      schema:
        type: string
        enum: [create, reject]
...
    LimitParam:
      name: limit
      in: query
      description: Maximum rows to return (default 100, capped at 1000).
      schema:
        type: integer
        default: 100
        maximum: 1000
    OffsetParam:
      name: offset
      in: query
      description: Zero-based row offset for server-side pagination.
      schema:
        type: integer
        default: 0
        minimum: 0
    PageParam:
      name: page
      in: query
      description: >
        One-based page number for server-side pagination. When present, the
        backend computes offset as `(page - 1) * limit`.
      schema:
        type: integer
        minimum: 1

  # ---------------------------------------------------------------------------
  # Responses
  # ---------------------------------------------------------------------------
  responses:
    MalformedRequestError:
      description: Malformed HTTP request or JSON syntax.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    ValidationError:
      description: Syntactically valid request failed schema or parameter validation.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
    ConflictError:
      description: >-
        Resource conflict, including an existing code, protected runtime state,
        or a configured dependency that prevents the mutation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    AccountMissingOrNotFoundError:
      description: >
        Either the addressed resource does not exist (`not_found`) or the
        request named an unknown account and asked to reject it
        (`account_missing`, with the account code in the envelope).
      content:
        application/json:
          schema:
            anyOf:
              - $ref: "#/components/schemas/Error"
              - $ref: "#/components/schemas/AccountMissingError"
...
    InternalError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    Id:
      type: string
      description: >
        Opaque public handle. No surrogate or engine id is ever exposed.
...
    ProblemDetails:
      type: object
      additionalProperties: false
      required: [type, title, status, detail, errors]
      properties:
        type:
          type: string
          description: >
            RFC 9457 problem type URI. Officer uses `about:blank` for generic
            HTTP errors.
          enum: ["about:blank"]
        title:
          type: string
          description: Canonical HTTP status title.
        status:
          type: integer
          enum: [400, 422]
        detail:
          type: string
          description: Human-readable explanation for this occurrence.
        errors:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ProblemError"
    ProblemError:
      type: object
      additionalProperties: false
      required: [code, pointer]
      properties:
        code:
          type: string
          description: Stable machine-readable error category.
        constraint:
          type: string
          description: Violated request constraint, when known.
        pointer:
          type: string
          description: JSON Pointer to the invalid member; empty means the request root.
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: >
                Machine-readable code: `validation`, `not_found`,
                `account_missing`, `conflict`, `currency_change_blocked`, `has_dependents`,
                `terminal_order`, `execution_report_required`, `precondition`,
                `too_large`, `forbidden`, `reserved_group`,
                `engine_restarting`, `signing`, `upstream`, `not_implemented`,
                or `internal`.
              example: not_found
            message:
              type: string
              example: account not found
...
    AccountMissingError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, account]
          properties:
            code:
              type: string
              enum: [account_missing]
            message:
              type: string
              example: account "acc-1" does not exist
            account:
              type: string
              description: The account code the request named.
              example: acc-1

    # Business CSV
...
    Balance:
      type: object
      required: [account, asset, available, held, incoming, realizedPnl, realizedPnlHaltReason, averageEntryPrice, accountCurrency, updatedAt]
      properties:
        account:
          type: string
        asset:
          type: string
          description: Balance asset symbol.
        available:
          type: string
          description: Exact decimal string. A quantity of `asset`.
        held:
          type: string
          description: Exact decimal string. A quantity of `asset`.
        incoming:
          type: string
          description: Exact decimal string. A quantity of `asset`.
        realizedPnl:
          type: string
          description: Exact decimal string, denominated in `accountCurrency`.
        realizedPnlHaltReason:
          type: string
          enum: ["", missing_fx, missing_account_currency, missing_initial_pnl, missing_cost_basis, arithmetic_overflow]
          description: Engine-reported reason position P&L calculation is halted; empty when the P&L value is current.
        averageEntryPrice:
          type: string
          description: Exact decimal string, denominated in `accountCurrency`.
        accountCurrency:
          type: string
          description: >-
            Asset code denominating `realizedPnl` and `averageEntryPrice`: the
            account's effective currency, resolved from the account, its group,
            and the reserved default group. One (account, asset) slot merges
            fills against every quote asset the asset trades against, so a
            single fill's quote asset does not denominate these values. Empty
            when no tier of the cascade is set, which leaves both values
            without a unit.
        updatedAt:
          type: string
          format: date-time
    BalanceRealizedPnlRequest:
      type: object
      additionalProperties: false
      required: [asset, realizedPnl]
      properties:
        asset:
          type: string
          description: Asset whose per-(account, asset) balance row is updated.
        realizedPnl:
          type: string
          description: Current realized P&L as an exact decimal string.

    # Adjustment
    AdjustmentAmount:
      type: object
      required: [mode, value]
      properties:
        mode:
          type: string
          description: Adjustment mode (e.g. `set`, `delta`).
        value:
          type: string
          description: Exact decimal string.
    AdjustmentBounds:
      type: object
      properties:
        lower:
          type: string
          description: Inclusive lower bound - exact decimal string.
        upper:
          type: string
          description: Inclusive upper bound - exact decimal string.
    AdjustmentRequest:
      type: object
      additionalProperties: false
      required: [asset]
      properties:
        id:
          $ref: "#/components/schemas/Id"
          description: >
            Optional caller-supplied public id for the adjustment record.
            Write-only on create: when omitted the server generates one and
            returns it under the record's `id`. A duplicate is a 409.
            Never a surrogate id.
        asset:
          type: string
          description: Asset whose spot-funds balances are adjusted.
        averageEntryPrice:
          type: string
          description: Exact decimal string.
        realizedPnl:
          type: string
          description: >
            Optional exact decimal snapshot for persisted cumulative realized
            P&L on the adjusted asset. This is recorded with the adjustment
            history and is not sent to the SDK account-adjustment API.
        balance:
          description: Requested available-balance change or target.
          $ref: "#/components/schemas/AdjustmentAmount"
        balanceBounds:
          description: Optional guardrail for the resulting available balance.
          $ref: "#/components/schemas/AdjustmentBounds"
        held:
          description: Requested held-balance change or target.
          $ref: "#/components/schemas/AdjustmentAmount"
        heldBounds:
          description: Optional guardrail for the resulting held balance.
          $ref: "#/components/schemas/AdjustmentBounds"
        incoming:
          description: Requested incoming-balance change or target.
          $ref: "#/components/schemas/AdjustmentAmount"
        incomingBounds:
          description: Optional guardrail for the resulting incoming balance.
          $ref: "#/components/schemas/AdjustmentBounds"
    AdjustmentAccepted:
      type: object
      required: [balanceDelta, balanceResult, heldDelta, heldResult, incomingDelta, incomingResult, realizedPnlResult]
      properties:
        balanceDelta:
          type: string
          description: Exact decimal string.
        balanceResult:
          type: string
          description: Exact decimal string.
        heldDelta:
          type: string
          description: Exact decimal string.
        heldResult:
          type: string
          description: Exact decimal string.
        incomingDelta:
          type: string
          description: Exact decimal string.
        incomingResult:
          type: string
          description: Exact decimal string.
        realizedPnlResult:
          type: object
          required: [delta, result]
          properties:
            delta:
              type: string
              description: Exact decimal string.
            result:
              type: string
              description: Exact decimal string.
        realizedPnlHaltReason:
          type: string
          enum: ["", missing_fx, missing_account_currency, missing_initial_pnl, missing_cost_basis, arithmetic_overflow]
          description: Engine-reported reason position P&L calculation is halted.
        averageEntryPrice:
          type: string
          description: Engine-reported average entry price when present.
...
    AdjustmentRejected:
      type: object
      required: [code, reason]
      properties:
        code:
          type: string
        scope:
          type: string
        policy:
          type: string
        reason:
          type: string
        details:
          type: string
    AdjustmentOutcome:
      type: object
      description: Exactly one of `accepted` or `rejected` is present.
      properties:
        accepted:
          $ref: "#/components/schemas/AdjustmentAccepted"
        rejected:
          $ref: "#/components/schemas/AdjustmentRejected"
    Adjustment:
      type: object
      required: [id, at, account, source, status, request, outcome]
      properties:
        id:
          description: The adjustment's opaque public handle.
          $ref: "#/components/schemas/Id"
        at:
          type: string
          format: date-time
        account:
          type: string
          description: Account code the adjustment applied to.
        principal:
          type: string
          description: Principal that performed the adjustment; omitted when absent.
        source:
          type: string
          description: >
            Originating surface: `panel`, `api`, `mcp`, or `system`.
          enum: [panel, api, mcp, system]
        status:
          type: string
          description: "`accepted` or `rejected`."
          enum: [accepted, rejected]
        request:
          $ref: "#/components/schemas/AdjustmentRequest"
        outcome:
          $ref: "#/components/schemas/AdjustmentOutcome"

    # Order
...